Foundation: Postgres schema, compose file, test scaffold (#42)
- Fresh EF Core migration replaces InitAppSchema: StreamId is bigint
  (was int), actor_user_links is dropped, auth_users is narrowed to
  username/password hash/user id/must-change-password flag, and
  rank_thresholds seed data matches InMemoryRankThresholdProvider's
  canonical values.
- AppDbContext and its row classes updated to match: ActorUserLinkRow
  removed, AuthUserRow narrowed (also widened UserId to long, since it
  holds the same ulong Discord snowflake IDs StreamId does).
- AppDbContext and NpgsqlDataSource registered in Program.cs's DI
  composition root, reading the RngHelpdeskDB connection string.
- Added a design-time IDesignTimeDbContextFactory so `dotnet ef`
  commands work without a running host.
- Added a Podman-compatible docker-compose.yml for local Postgres.
- Added RngHelpdesk.Infrastructure.Tests (xunit + Testcontainers.PostgreSql),
  wired into RngHelpdesk.slnx, with a migration test proving the schema
  applies cleanly to a real ephemeral Postgres and matches expectations
  (schemas/tables, bigint StreamId, seeded rank thresholds).

+++ RngHelpdesk.Infrastructure/Persistence/Contexts/EventStreamRow.cs
@@ -3,7 +3,7 @@
 internal sealed class EventStreamRow
 {
     public string StreamType { get; set; } = default!;
-    public int StreamId { get; set; }
+    public long StreamId { get; set; }
 
     public int CurrentVersion { get; set; }
